refactor: move shot streak helper
